[FIX] Moved parse config to utils
